refactor(example): give each settings section its own widget
The panel's five sections were private build methods on one 1,500-line class.
Each is now a widget in its own file, taking the settings it edits and the
callback it reports through — and nothing else. Only the data section needed
more than that pair, for the generate button it owns.

The panel is left with what a panel is: a fixed-width scrolling column that
composes a header, five sections, and the performance monitor. 1,526 lines
became 178.

Pure move, and the characterisation test says so: the same five section titles,
the same 25 switches, 12 dropdowns and 20 sliders, without a line of the test
changing.

Closes #59
