+++ example/lib/pages/playground/widgets/settings_panel.dart
@@ -1,9 +1,12 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_table_plus/flutter_table_plus.dart';
 
 import '../models/playground_settings.dart';
 import 'performance_monitor.dart';
-import 'settings_controls.dart';
+import 'sections/data_settings_section.dart';
+import 'sections/feature_toggles_section.dart';
+import 'sections/header_border_section.dart';
+import 'sections/style_settings_section.dart';
+import 'sections/tooltip_settings_section.dart';
 
 /// Settings panel widget for the playground
 ///
@@ -57,7 +60,12 @@ class SettingsPanel extends StatelessWidget {
             const SizedBox(height: 24),
 
             // Data Settings
-            _buildDataSettings(),
+            DataSettingsSection(
+              settings: settings,
+              onSettingsChanged: onSettingsChanged,
+              onGenerateData: onGenerateData,
+              isGenerating: isGenerating,
+            ),
             const SizedBox(height: 16),
 
             // Width persistence demo buttons
@@ -102,19 +110,32 @@ class SettingsPanel extends StatelessWidget {
             const SizedBox(height: 24),
 
             // Style Settings
-            _buildStyleSettings(),
+            StyleSettingsSection(
+              settings: settings,
+              onSettingsChanged: onSettingsChanged,
+              onRandomizeWidths: onRandomizeWidths,
+            ),
             const SizedBox(height: 24),
 
             // Header Border/Divider Settings
-            _buildHeaderBorderSettings(),
+            HeaderBorderSection(
+              settings: settings,
+              onSettingsChanged: onSettingsChanged,
+            ),
             const SizedBox(height: 24),
 
             // Feature Toggles
-            _buildFeatureToggles(),
+            FeatureTogglesSection(
+              settings: settings,
+              onSettingsChanged: onSettingsChanged,
+            ),
             const SizedBox(height: 24),
 
             // Tooltip Settings
-            _buildTooltipSettings(),
+            TooltipSettingsSection(
+              settings: settings,
+              onSettingsChanged: onSettingsChanged,
+            ),
             const SizedBox(height: 24),
 
             // Performance Monitor
@@ -152,898 +173,6 @@ class SettingsPanel extends StatelessWidget {
           ),
         ),
       ],
-    );
-  }
-
-  Widget _buildDataSettings() {
-    return buildSection(
-      title: 'Data Settings',
-      icon: Icons.data_array,
-      color: Colors.green.shade700,
-      borderColor: Colors.green.shade200,
-      initiallyExpanded: true,
-      children: [
-        const SizedBox(height: 12),
-
-        // Row count label
-        Row(
-          mainAxisAlignment: MainAxisAlignment.spaceBetween,
-          children: [
-            const Text(
-              'Row Count',
-              style: TextStyle(
-                fontSize: 14,
-                fontWeight: FontWeight.w500,
-              ),
-            ),
-            Container(
-              padding: const EdgeInsets.symmetric(horizontal: 10, vertical: 4),
-              decoration: BoxDecoration(
-                color: Colors.green.shade100,
-                borderRadius: BorderRadius.circular(4),
-              ),
-              child: Text(
-                formatNumber(settings.rowCount),
-                style: TextStyle(
-                  fontSize: 14,
-                  fontWeight: FontWeight.bold,
-                  color: Colors.green.shade800,
-                ),
-              ),
-            ),
-          ],
-        ),
-        const SizedBox(height: 12),
-
-        // Logarithmic slider (5 to 100,000)
-        buildLogSlider(
-          value: settings.rowCount.toDouble(),
-          min: 5,
-          max: 100000,
-          onChanged: (value) {
-            onSettingsChanged(settings.copyWith(rowCount: value.round()));
-          },
-        ),
-        const SizedBox(height: 16),
-
-        // Quick select buttons. The "5" preset leaves a large empty area
-        // below the last row, useful for testing drag-selection edge cases.
-        Wrap(
-          spacing: 8,
-          runSpacing: 8,
-          children: [
-            _buildQuickButton('5', 5),
-            _buildQuickButton('100', 100),
-            _buildQuickButton('1K', 1000),
-            _buildQuickButton('10K', 10000),
-            _buildQuickButton('100K', 100000),
-          ],
-        ),
-        const SizedBox(height: 16),
-
-        // Generate button
-        SizedBox(
-          width: double.infinity,
-          height: 48,
-          child: ElevatedButton.icon(
-            onPressed: isGenerating ? null : onGenerateData,
-            icon: isGenerating
-                ? const SizedBox(
-                    width: 20,
-                    height: 20,
-                    child: CircularProgressIndicator(strokeWidth: 2),
-                  )
-                : const Icon(Icons.refresh),
-            label: Text(isGenerating ? 'Generating...' : 'Generate Data'),
-            style: ElevatedButton.styleFrom(
-              backgroundColor: Colors.green.shade600,
-              foregroundColor: Colors.white,
-              textStyle: const TextStyle(
-                fontSize: 16,
-                fontWeight: FontWeight.w600,
-              ),
-            ),
-          ),
-        ),
-      ],
-    );
-  }
-
-  Widget _buildStyleSettings() {
-    return buildSection(
-      title: 'Style Settings',
-      icon: Icons.palette,
-      color: Colors.purple.shade700,
-      borderColor: Colors.purple.shade200,
-      children: [
-        const SizedBox(height: 12),
-
-        // Font family
-        buildDropdownRow<String>(
-          label: 'Font Family',
-          value: settings.fontFamily,
-          items: const [
-            'default',
-            'pretendard',
-            'notoSansKr',
-            'inter',
-            'firaCode'
-          ],
-          itemLabel: (font) => switch (font) {
-            'default' => 'Default (Roboto)',
-            'pretendard' => 'Pretendard',
-            'notoSansKr' => 'Noto Sans KR',
-            'inter' => 'Inter',
-            'firaCode' => 'Fira Code',
-            _ => font,
-          },
-          onChanged: (value) {
-            onSettingsChanged(settings.copyWith(fontFamily: value));
-          },
-        ),
-        const SizedBox(height: 16),
-
-        // Scale (zoom)
-        buildSliderSetting(
-          label: 'Scale',
-          value: settings.scale,
-          min: 0.25,
-          max: 3.0,
-          unit: 'x',
-          decimalPlaces: 2,
-          onChanged: (value) {
-            onSettingsChanged(settings.copyWith(scale: value));
-          },
-        ),
-        Align(
-          alignment: Alignment.centerRight,
-          child: TextButton(
-            onPressed: () {
-              onSettingsChanged(settings.copyWith(scale: 1.0));
-            },
-            style: TextButton.styleFrom(
-              foregroundColor: Colors.purple.shade700,
-              padding: const EdgeInsets.symmetric(horizontal: 8),
-              minimumSize: const Size(0, 28),
-            ),
-            child: const Text('Reset (1.0x)', style: TextStyle(fontSize: 11)),
-          ),
-        ),
-        buildSwitchTile(
-          label: 'Block Modifier+Scroll',
-          value: settings.blockModifierScroll,
-          onChanged: (value) {
-            onSettingsChanged(settings.copyWith(blockModifierScroll: value));
-          },
-        ),
-        const SizedBox(height: 12),
-
-        // Column min width
-        buildSliderSetting(
-          label: 'Col Min Width',
-          value: settings.columnMinWidth,
-          min: 20,
-          max: 150,
-          unit: 'px',
-          onChanged: (value) {
-            onSettingsChanged(settings.copyWith(columnMinWidth: value));
-          },
-        ),
-        const SizedBox(height: 12),
-
-        // Randomize column widths button
-        if (onRandomizeWidths != null)
-          SizedBox(
-            width: double.infinity,
-            child: OutlinedButton.icon(
-              onPressed: onRandomizeWidths,
-              icon: const Icon(Icons.shuffle, size: 18),
-              label: const Text('Randomize Column Widths'),
-              style: OutlinedButton.styleFrom(
-                foregroundColor: Colors.purple.shade700,
-                side: BorderSide(color: Colors.purple.shade300),
-              ),
-            ),
-          ),
-        const SizedBox(height: 16),
-
-        // Row height
-        buildSliderSetting(
-          label: 'Row Height',
-          value: settings.rowHeight,
-          min: 30,
-          max: 100,
-          unit: 'px',
-          onChanged: (value) {
-            onSettingsChanged(settings.copyWith(rowHeight: value));
-          },
-        ),
-        const SizedBox(height: 16),
-
-        // Font size
-        buildSliderSetting(
-          label: 'Font Size',
-          value: settings.fontSize,
-          min: 10,
-          max: 24,
-          unit: 'px',
-          onChanged: (value) {
-            onSettingsChanged(settings.copyWith(fontSize: value));
-          },
-        ),
-        const SizedBox(height: 16),
-
-        // Horizontal padding
-        buildSliderSetting(
-          label: 'H-Padding',
-          value: settings.horizontalPadding,
-          min: 4,
-          max: 32,
-          unit: 'px',
-          onChanged: (value) {
-            onSettingsChanged(settings.copyWith(horizontalPadding: value));
-          },
-        ),
-        const SizedBox(height: 16),
-
-        // Vertical padding
-        buildSliderSetting(
-          label: 'V-Padding',
-          value: settings.verticalPadding,
-          min: 4,
-          max: 24,
-          unit: 'px',
-          onChanged: (value) {
-            onSettingsChanged(settings.copyWith(verticalPadding: value));
-          },
-        ),
-        const SizedBox(height: 16),
-
-        // Checkbox tap target size
-        buildSliderSetting(
-          label: 'Checkbox Tap Size',
-          value: settings.checkboxTapTargetSize,
-          min: 18,
-          max: 56,
-          unit: 'px',
-          onChanged: (value) {
-            onSettingsChanged(settings.copyWith(checkboxTapTargetSize: value));
-          },
-        ),
-        const SizedBox(height: 16),
-
-        // Sort icon width
-        buildSliderSetting(
-          label: 'Sort Icon Width',
-          value: settings.sortIconWidth,
-          min: 8,
-          max: 32,
-          unit: 'px',
-          onChanged: (value) {
-            onSettingsChanged(settings.copyWith(sortIconWidth: value));
-          },
-        ),
-      ],
-    );
-  }
-
-  Widget _buildHeaderBorderSettings() {
-    return buildSection(
-      title: 'Header Border / Divider',
-      icon: Icons.border_all,
-      color: Colors.teal.shade700,
-      borderColor: Colors.teal.shade200,
-      children: [
-        // --- Top Border ---
-        buildSwitchTile(
-          label: 'Top Border',
-          value: settings.headerTopBorderShow,
-          onChanged: (value) {
-            onSettingsChanged(settings.copyWith(headerTopBorderShow: value));
-          },
-        ),
-        if (settings.headerTopBorderShow)
-          Padding(
-            padding: const EdgeInsets.only(left: 8),
-            child: buildSliderSetting(
-              label: 'Thickness',
-              value: settings.headerTopBorderThickness,
-              min: 0.5,
-              max: 6,
-              unit: 'px',
-              onChanged: (value) {
-                onSettingsChanged(
-                    settings.copyWith(headerTopBorderThickness: value));
-              },
-            ),
-          ),
-        const SizedBox(height: 4),
-
-        // --- Bottom Border ---
-        buildSwitchTile(
-          label: 'Bottom Border',
-          value: settings.headerBottomBorderShow,
-          onChanged: (value) {
-            onSettingsChanged(settings.copyWith(headerBottomBorderShow: value));
-          },
-        ),
-        if (settings.headerBottomBorderShow)
-          Padding(
-            padding: const EdgeInsets.only(left: 8),
-            child: buildSliderSetting(
-              label: 'Thickness',
-              value: settings.headerBottomBorderThickness,
-              min: 0.5,
-              max: 6,
-              unit: 'px',
-              onChanged: (value) {
-                onSettingsChanged(
-                    settings.copyWith(headerBottomBorderThickness: value));
-              },
-            ),
-          ),
-
-        const Divider(height: 24),
-
-        // --- Vertical Divider ---
-        buildSwitchTile(
-          label: 'Vertical Divider',
-          value: settings.headerVerticalDividerShow,
-          onChanged: (value) {
-            onSettingsChanged(
-                settings.copyWith(headerVerticalDividerShow: value));
-          },
-        ),
-        if (settings.headerVerticalDividerShow) ...[
-          Padding(
-            padding: const EdgeInsets.only(left: 8),
-            child: buildSliderSetting(
-              label: 'Thickness',
-              value: settings.headerVerticalDividerThickness,
-              min: 0.5,
-              max: 6,
-              unit: 'px',
-              onChanged: (value) {
-                onSettingsChanged(
-                    settings.copyWith(headerVerticalDividerThickness: value));
-              },
-            ),
-          ),
-          const SizedBox(height: 8),
-          Padding(
-            padding: const EdgeInsets.only(left: 8),
-            child: buildSliderSetting(
-              label: 'Indent (top)',
-              value: settings.headerVerticalDividerIndent,
-              min: 0,
-              max: 24,
-              unit: 'px',
-              onChanged: (value) {
-                onSettingsChanged(
-                    settings.copyWith(headerVerticalDividerIndent: value));
-              },
-            ),
-          ),
-          const SizedBox(height: 8),
-          Padding(
-            padding: const EdgeInsets.only(left: 8),
-            child: buildSliderSetting(
-              label: 'End Indent (bottom)',
-              value: settings.headerVerticalDividerEndIndent,
-              min: 0,
-              max: 24,
-              unit: 'px',
-              onChanged: (value) {
-                onSettingsChanged(
-                    settings.copyWith(headerVerticalDividerEndIndent: value));
-              },
-            ),
-          ),
-        ],
-
-        const Divider(height: 24),
-
-        // --- Resize Handle ---
-        if (settings.resizableEnabled) ...[
-          buildSliderSetting(
-            label: 'Handle Hit Width',
-            value: settings.resizeHandleWidth,
-            min: 4,
-            max: 24,
-            unit: 'px',
-            onChanged: (value) {
-              onSettingsChanged(settings.copyWith(resizeHandleWidth: value));
-            },
-          ),
-          const SizedBox(height: 8),
-          buildSliderSetting(
-            label: 'Handle Thickness',
-            value: settings.resizeHandleThickness,
-            min: 1,
-            max: 6,
-            unit: 'px',
-            onChanged: (value) {
-              onSettingsChanged(
-                  settings.copyWith(resizeHandleThickness: value));
-            },
-          ),
-          const SizedBox(height: 8),
-          buildSliderSetting(
-            label: 'Handle Indent (top)',
-            value: settings.resizeHandleIndent,
-            min: 0,
-            max: 24,
-            unit: 'px',
-            onChanged: (value) {
-              onSettingsChanged(settings.copyWith(resizeHandleIndent: value));
-            },
-          ),
-          const SizedBox(height: 8),
-          buildSliderSetting(
-            label: 'Handle End Indent',
-            value: settings.resizeHandleEndIndent,
-            min: 0,
-            max: 24,
-            unit: 'px',
-            onChanged: (value) {
-              onSettingsChanged(
-                  settings.copyWith(resizeHandleEndIndent: value));
-            },
-          ),
-        ],
-      ],
-    );
-  }
-
-  Widget _buildFeatureToggles() {
-    return buildSection(
-      title: 'Feature Toggles',
-      icon: Icons.toggle_on,
-      color: Colors.orange.shade700,
-      borderColor: Colors.orange.shade200,
-      children: [
-        const SizedBox(height: 12),
-
-        buildSwitchTile(
-          label: 'Sorting',
-          value: settings.sortingEnabled,
-          onChanged: (value) {
-            onSettingsChanged(settings.copyWith(sortingEnabled: value));
-          },
-        ),
-
-        buildSwitchTile(
-          label: 'Editing',
-          value: settings.editingEnabled,
-          onChanged: (value) {
-            onSettingsChanged(settings.copyWith(editingEnabled: value));
-          },
-        ),
-
-        buildSwitchTile(
-          label: 'Column Reorder',
-          value: settings.columnReorderEnabled,
-          onChanged: (value) {
-            onSettingsChanged(settings.copyWith(columnReorderEnabled: value));
-          },
-        ),
-
-        buildSwitchTile(
-          label: 'Column Resize',
-          value: settings.resizableEnabled,
-          onChanged: (value) {
-            onSettingsChanged(settings.copyWith(resizableEnabled: value));
-          },
-        ),
-
-        buildSwitchTile(
-          label: 'Stretch Last Column',
-          value: settings.stretchLastColumn,
-          onChanged: (value) {
-            onSettingsChanged(settings.copyWith(stretchLastColumn: value));
-          },
-        ),
-
-        buildSwitchTile(
-          label: 'Alternate Rows',
-          value: settings.showAlternateRows,
-          onChanged: (value) {
-            onSettingsChanged(settings.copyWith(showAlternateRows: value));
-          },
-        ),
-
-        buildSwitchTile(
-          label: 'Row Card Tooltip',
-          value: settings.rowCardTooltip,
-          onChanged: (value) {
-            onSettingsChanged(settings.copyWith(rowCardTooltip: value));
-          },
-        ),
-
-        buildDropdownRow<InkColorOption>(
-          label: 'Splash Color',
-          value: settings.splashColor,
-          items: InkColorOption.values,
-          itemLabel: (option) => option.label,
-          onChanged: (option) {
-            onSettingsChanged(settings.copyWith(splashColor: option));
-          },
-        ),
-
-        buildDropdownRow<InkColorOption>(
-          label: 'Hover Color',
-          value: settings.hoverColor,
-          items: InkColorOption.values,
-          itemLabel: (option) => option.label,
-          onChanged: (option) {
-            onSettingsChanged(settings.copyWith(hoverColor: option));
-          },
-        ),
-
-        buildDropdownRow<InkColorOption>(
-          label: 'Highlight Color',
-          value: settings.highlightColor,
-          items: InkColorOption.values,
-          itemLabel: (option) => option.label,
-          onChanged: (option) {
-            onSettingsChanged(settings.copyWith(highlightColor: option));
-          },
-        ),
-
-        const Padding(
-          padding: EdgeInsets.symmetric(vertical: 4),
-          child: Text(
-            'Row splash needs Selection on and Editing off — while editing, the '
-            'ink layer stays for row gestures but the selection splash is '
-            'suppressed.',
-            style: TextStyle(fontSize: 11, color: Colors.black54),
-          ),
-        ),
-
-        buildSwitchTile(
-          label: 'Show Dividers',
-          value: settings.showDividers,
-          onChanged: (value) {
-            onSettingsChanged(settings.copyWith(showDividers: value));
-          },
-        ),
-
-        buildSwitchTile(
-          label: 'Merged Rows',
-          value: settings.mergedRowsEnabled,
-          onChanged: (value) {
-            onSettingsChanged(settings.copyWith(mergedRowsEnabled: value));
-          },
-        ),
-
-        buildSwitchTile(
-          label: 'Dynamic Row Height',
-          value: settings.dynamicRowHeight,
-          onChanged: (value) {
-            onSettingsChanged(settings.copyWith(dynamicRowHeight: value));
-          },
-        ),
-
-        buildSwitchTile(
-          label: 'Dim Inactive Rows',
-          value: settings.dimInactiveRows,
-          onChanged: (value) {
-            onSettingsChanged(settings.copyWith(dimInactiveRows: value));
-          },
-        ),
-
-        buildSwitchTile(
-          label: 'Selection (Checkbox)',
-          value: settings.selectionEnabled,
-          onChanged: (value) {
-            onSettingsChanged(settings.copyWith(selectionEnabled: value));
-          },
-        ),
-
-        if (settings.selectionEnabled) ...[
-          buildSwitchTile(
-            label: 'Show Checkbox Column',
-            value: settings.showCheckboxColumn,
-            onChanged: (value) {
-              onSettingsChanged(settings.copyWith(showCheckboxColumn: value));
-            },
-          ),
-          buildSwitchTile(
-            label: 'Select All',
-            value: settings.selectAllEnabled,
-            onChanged: (value) {
-              onSettingsChanged(settings.copyWith(selectAllEnabled: value));
-            },
-          ),
-          buildSwitchTile(
-            label: 'Drag Selection',
-            value: settings.dragSelectionEnabled,
-            onChanged: (value) {
-              onSettingsChanged(settings.copyWith(dragSelectionEnabled: value));
-            },
-          ),
-          buildSwitchTile(
-            label: 'Cell Tap Toggles Checkbox',
-            value: settings.cellTapTogglesCheckbox,
-            onChanged: (value) {
-              onSettingsChanged(
-                  settings.copyWith(cellTapTogglesCheckbox: value));
-            },
-          ),
-          buildSwitchTile(
-            label: 'Show Row Checkbox',
-            value: settings.showRowCheckbox,
-            onChanged: (value) {
-              onSettingsChanged(settings.copyWith(showRowCheckbox: value));
-            },
-          ),
-        ],
-
-        const SizedBox(height: 12),
-        const Divider(),
-        const SizedBox(height: 12),
-
-        // Selection mode dropdown
-        buildDropdownRow<SelectionMode>(
-          label: 'Selection Mode',
-          value: settings.selectionMode,
-          items: SelectionMode.values,
-          itemLabel: (mode) => mode.name.toUpperCase(),
-          onChanged: (value) {
-            onSettingsChanged(settings.copyWith(selectionMode: value));
-          },
-        ),
-        const SizedBox(height: 8),
-
-        // Sort cycle order dropdown
-        buildDropdownRow<SortCycleOrder>(
-          label: 'Sort Cycle',
-          value: settings.sortCycleOrder,
-          items: SortCycleOrder.values,
-          itemLabel: (order) => order == SortCycleOrder.ascendingFirst
-              ? 'ASC First'
-              : 'DESC First',
-          onChanged: (value) {
-            onSettingsChanged(settings.copyWith(sortCycleOrder: value));
-          },
-        ),
-      ],
-    );
-  }
-
-  Widget _buildTooltipSettings() {
-    return buildSection(
-      title: 'Tooltip Settings',
-      icon: Icons.chat_bubble_outline,
-      color: Colors.indigo.shade700,
-      borderColor: Colors.indigo.shade200,
-      children: [
-        const SizedBox(height: 12),
-
-        // Tooltip enabled toggle
-        buildSwitchTile(
-          label: 'Tooltip Enabled',
-          value: settings.tooltipEnabled,
-          onChanged: (value) {
-            onSettingsChanged(settings.copyWith(tooltipEnabled: value));
-          },
-        ),
-
-        if (settings.tooltipEnabled) ...[
-          const SizedBox(height: 8),
-
-          // Cell tooltip behavior
-          buildDropdownRow<TooltipBehavior>(
-            label: 'Cell Tooltip',
-            value: settings.tooltipBehavior,
-            items: TooltipBehavior.values,
-            itemLabel: (behavior) => switch (behavior) {
-              TooltipBehavior.always => 'Always',
-              TooltipBehavior.never => 'Never',
-              TooltipBehavior.onlyTextOverflow => 'On Overflow',
-            },
-            onChanged: (value) {
-              onSettingsChanged(settings.copyWith(tooltipBehavior: value));
-            },
-          ),
-          const SizedBox(height: 4),
-
-          // Header tooltip behavior
-          buildDropdownRow<TooltipBehavior>(
-            label: 'Header Tooltip',
-            value: settings.headerTooltipBehavior,
-            items: TooltipBehavior.values,
-            itemLabel: (behavior) => switch (behavior) {
-              TooltipBehavior.always => 'Always',
-              TooltipBehavior.never => 'Never',
-              TooltipBehavior.onlyTextOverflow => 'On Overflow',
-            },
-            onChanged: (value) {
-              onSettingsChanged(
-                  settings.copyWith(headerTooltipBehavior: value));
-            },
-          ),
-
-          const Divider(height: 24),
-
-          // Wait duration slider
-          buildSliderSetting(
-            label: 'Wait Duration',
-            value: settings.tooltipWaitDurationMs.toDouble(),
-            min: 0,
-            max: 2000,
-            unit: 'ms',
-            onChanged: (value) {
-              onSettingsChanged(
-                  settings.copyWith(tooltipWaitDurationMs: value.round()));
-            },
-          ),
-
-          const Divider(height: 24),
-
-          // Direction
-          buildDropdownRow<TooltipDirection>(
-            label: 'Direction',
-            value: settings.tooltipDirection,
-            items: TooltipDirection.values,
-            itemLabel: (d) => switch (d) {
-              TooltipDirection.top => 'Top',
-              TooltipDirection.bottom => 'Bottom',
-              TooltipDirection.left => 'Left',
-              TooltipDirection.right => 'Right',
-            },
-            onChanged: (value) {
-              onSettingsChanged(settings.copyWith(tooltipDirection: value));
-            },
-          ),
-          const SizedBox(height: 4),
-
-          // Anchor — cells
-          buildDropdownRow<TooltipAnchor>(
-            label: 'Cell Anchor',
-            value: settings.tooltipAnchor,
-            items: TooltipAnchor.values,
-            itemLabel: (a) => switch (a) {
-              TooltipAnchor.child => 'Child',
-              TooltipAnchor.pointer => 'Pointer',
-            },
-            onChanged: (value) {
-              onSettingsChanged(settings.copyWith(tooltipAnchor: value));
-            },
-          ),
-          Padding(
-            padding: const EdgeInsets.only(left: 8, bottom: 4),
-            child: Text(
-              'The row card always anchors at the pointer; it ignores this',
-              style: TextStyle(fontSize: 11, color: Colors.grey.shade600),
-            ),
-          ),
-
-          // Anchor — headers
-          buildDropdownRow<HeaderTooltipAnchor>(
-            label: 'Header Anchor',
-            value: settings.headerTooltipAnchor,
-            items: HeaderTooltipAnchor.values,
-            itemLabel: (a) => switch (a) {
-              HeaderTooltipAnchor.followCells => 'Follow Cells',
-              HeaderTooltipAnchor.child => 'Child',
-              HeaderTooltipAnchor.pointer => 'Pointer',
-            },
-            onChanged: (value) {
-              onSettingsChanged(settings.copyWith(headerTooltipAnchor: value));
-            },
-          ),
-          const SizedBox(height: 4),
-
-          // Alignment
-          buildDropdownRow<TooltipAlignment>(
-            label: 'Alignment',
-            value: settings.tooltipAlignment,
-            items: TooltipAlignment.values,
-            itemLabel: (a) => switch (a) {
-              TooltipAlignment.start => 'Start',
-              TooltipAlignment.center => 'Center',
-              TooltipAlignment.end => 'End',
-              TooltipAlignment.startTargetCenter => 'Start Target Center',
-              TooltipAlignment.endTargetCenter => 'End Target Center',
-            },
-            onChanged: (value) {
-              onSettingsChanged(settings.copyWith(tooltipAlignment: value));
-            },
-          ),
-          Padding(
-            padding: const EdgeInsets.only(left: 8, bottom: 4),
-            child: Text(
-              settings.tooltipAnchor == TooltipAnchor.pointer
-                  ? "Against the pointer there are no target edges, so this "
-                      "picks which of the tooltip's own edges lands on the cursor"
-                  : 'Picks which edge of the target the tooltip lines up with',
-              style: TextStyle(fontSize: 11, color: Colors.grey.shade600),
-            ),
-          ),
-
-          // Show Arrow
-          buildSwitchTile(
-            label: 'Show Arrow',
-            value: settings.tooltipShowArrow,
-            onChanged: (value) {
-              onSettingsChanged(settings.copyWith(tooltipShowArrow: value));
-            },
-          ),
-
-          // Offset slider
-          buildSliderSetting(
-            label: 'Offset',
-            value: settings.tooltipOffset,
-            min: 0,
-            max: 32,
-            unit: 'px',
-            onChanged: (value) {
-              onSettingsChanged(
-                  settings.copyWith(tooltipOffset: value.roundToDouble()));
-            },
-          ),
-
-          const Divider(height: 24),
-
-          // Demo: tooltipFormatter
-          buildSwitchTile(
-            label: 'tooltipFormatter (Email)',
-            value: settings.showTooltipFormatter,
-            onChanged: (value) {
-              onSettingsChanged(settings.copyWith(showTooltipFormatter: value));
-            },
-          ),
-          if (settings.showTooltipFormatter)
-            Padding(
-              padding: const EdgeInsets.only(left: 8, bottom: 4),
-              child: Text(
-                'Email column shows "Send to: {email}" tooltip',
-                style: TextStyle(fontSize: 11, color: Colors.grey.shade600),
-              ),
-            ),
-
-          // Demo: tooltipBuilder
-          buildSwitchTile(
-            label: 'tooltipBuilder (Name)',
-            value: settings.showTooltipBuilder,
-            onChanged: (value) {
-              onSettingsChanged(settings.copyWith(showTooltipBuilder: value));
-            },
-          ),
-          if (settings.showTooltipBuilder)
-            Padding(
-              padding: const EdgeInsets.only(left: 8, bottom: 4),
-              child: Text(
-                'Name column shows rich widget tooltip with employee details',
-                style: TextStyle(fontSize: 11, color: Colors.grey.shade600),
-              ),
-            ),
-        ],
-      ],
-    );
-  }
-
-  Widget _buildQuickButton(String label, int value) {
-    final isSelected = settings.rowCount == value;
-    return ElevatedButton(
-      onPressed: () {
-        onSettingsChanged(settings.copyWith(rowCount: value));
-      },
-      style: ElevatedButton.styleFrom(
-        backgroundColor:
-            isSelected ? Colors.green.shade600 : Colors.grey.shade200,
-        foregroundColor: isSelected ? Colors.white : Colors.grey.shade700,
-        padding: const EdgeInsets.symmetric(horizontal: 16, vertical: 10),
-        minimumSize: Size.zero,
-      ),
-      child: Text(
-        label,
-        style: const TextStyle(
-          fontSize: 13,
-          fontWeight: FontWeight.w600,
-        ),
-      ),
     );
   }
 }
